Moved stuff around to make it part of the InfoProvider package
