+++ src/providers/info.rs
@@ -16,7 +16,7 @@ use std::collections::HashMap;
 pub trait InfoProvider {
     fn safe_info(&mut self, safe: &str) -> Result<SafeInfo>;
     fn token_info(&mut self, token: &str) -> Result<TokenInfo>;
-    fn raw_request(&mut self, url: &str) -> ApiResult<String>;
+    fn safe_app_info(&mut self, url: &str) -> Result<SafeAppInfo>;
 }
 
 pub struct DefaultInfoProvider<'p> {
@@ -42,6 +42,22 @@ pub struct SafeInfo {
     pub nonce: u64,
     pub threshold: u64,
     pub owners: Vec<String>,
+}
+
+#[derive(Serialize, Debug, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SafeAppInfo {
+    pub name: String,
+    pub url: String,
+    pub logo_url: String,
+}
+
+#[derive(Deserialize, Debug, PartialEq)]
+struct Manifest {
+    pub(super) name: String,
+    pub(super) description: String,
+    #[serde(rename(deserialize = "iconPath"))]
+    pub(super) icon_path: String,
 }
 
 #[derive(Deserialize, Clone, Debug)]
@@ -78,8 +94,17 @@ impl InfoProvider for DefaultInfoProvider<'_> {
         }
     }
 
-    fn raw_request(&mut self, url: &str) -> ApiResult<String> {
-        self.cache.request_cached(self.client, &url, long_cache())
+    fn safe_app_info(&mut self, url: &str) -> Result<SafeAppInfo> {
+        let manifest_url = format!("{}/manifest.json", url);
+        let manifest_json = self
+            .cache
+            .request_cached(self.client, &manifest_url, long_cache())?;
+        let manifest = serde_json::from_str::<Manifest>(&manifest_json)?;
+        Ok(SafeAppInfo {
+            name: manifest.name.to_owned(),
+            url: url.to_owned(),
+            logo_url: format!("{}/{}", url, manifest.icon_path),
+        })
     }
 }
